download button working; next: layout
